LPS-136878 Lazy fetch localization
