+++ modules/apps/change-tracking/change-tracking-web/src/main/resources/META-INF/resources/publications/js/ChangeTrackingRenderView.js
@@ -13,15 +13,126 @@
  */
 
 import ClayAlert from '@clayui/alert';
-import ClayButton from '@clayui/button';
-import {Align, ClayDropDownWithItems} from '@clayui/drop-down';
+import ClayButton, {ClayButtonWithIcon} from '@clayui/button';
+import ClayDropDown, {Align, ClayDropDownWithItems} from '@clayui/drop-down';
 import ClayIcon from '@clayui/icon';
+import ClayLabel from '@clayui/label';
+import ClayLayout from '@clayui/layout';
 import ClayLink from '@clayui/link';
 import ClayNavigationBar from '@clayui/navigation-bar';
 import {fetch} from 'frontend-js-web';
 import React, {useEffect, useState} from 'react';
 
-export default ({dataURL, getCache, languageId, spritemap, updateCache}) => {
+const LocalizationDropdown = ({
+	currentLocale,
+	defaultLocale,
+	locales,
+	setSelectedLocale,
+	spritemap,
+}) => {
+	const [active, setActive] = useState(false);
+
+	return (
+		<div className="autofit-col publications-localization">
+			<ClayDropDown
+				active={active}
+				onActiveChange={setActive}
+				trigger={
+					<ClayButton
+						displayType="secondary"
+						monospaced
+						onClick={() => setActive(!active)}
+					>
+						<span className="inline-item">
+							<ClayIcon
+								spritemap={spritemap}
+								symbol={currentLocale.symbol}
+							/>
+						</span>
+						<span className="btn-section">
+							{currentLocale.label}
+						</span>
+					</ClayButton>
+				}
+			>
+				<ClayDropDown.ItemList>
+					{locales
+						.sort((a, b) => {
+							if (a.label === defaultLocale.label) {
+								return -1;
+							}
+							else if (b.label === defaultLocale.label) {
+								return 1;
+							}
+
+							return 0;
+						})
+						.map((locale) => {
+							return (
+								<ClayDropDown.Item
+									key={locale.label}
+									onClick={() => {
+										setActive(false);
+										setSelectedLocale(locale);
+									}}
+								>
+									<ClayLayout.ContentRow containerElement="span">
+										<ClayLayout.ContentCol
+											containerElement="span"
+											expand
+										>
+											<ClayLayout.ContentSection>
+												<ClayIcon
+													className="inline-item inline-item-before"
+													spritemap={spritemap}
+													symbol={locale.symbol}
+												/>
+
+												{locale.label}
+											</ClayLayout.ContentSection>
+										</ClayLayout.ContentCol>
+										<ClayLayout.ContentCol containerElement="span">
+											<ClayLayout.ContentSection>
+												<ClayLabel
+													displayType={
+														locale.label ===
+														defaultLocale.label
+															? 'info'
+															: 'success'
+													}
+												>
+													{locale.label ===
+													defaultLocale.label
+														? Liferay.Language.get(
+																'default'
+														  )
+														: Liferay.Language.get(
+																'translated'
+														  )}
+												</ClayLabel>
+											</ClayLayout.ContentSection>
+										</ClayLayout.ContentCol>
+									</ClayLayout.ContentRow>
+								</ClayDropDown.Item>
+							);
+						})}
+				</ClayDropDown.ItemList>
+			</ClayDropDown>
+		</div>
+	);
+};
+
+export default ({
+	dataURL,
+	defaultLocale,
+	description,
+	discardURL,
+	getCache,
+	showDropdown,
+	spritemap,
+	title,
+	updateCache,
+}) => {
 	const CHANGE_TYPE_ADDED = 'added';
 	const CHANGE_TYPE_DELETED = 'deleted';
 	const CHANGE_TYPE_MODIFIED = 'modified';
@@ -34,6 +145,7 @@ export default ({dataURL, getCache, languageId, spritemap, updateCache}) => {
 	const VIEW_UNIFIED = 'VIEW_UNIFIED';
 
 	const [loading, setLoading] = useState(false);
+	const [selectedLocale, setSelectedLocale] = useState(defaultLocale);
 	const [state, setState] = useState({
 		contentType: CONTENT_TYPE_DISPLAY,
 		renderData: null,
@@ -191,6 +303,33 @@ export default ({dataURL, getCache, languageId, spritemap, updateCache}) => {
 			});
 	}, [dataURL, getCache, updateCache]);
 
+	let currentLocale = selectedLocale;
+	let currentTitle = title;
+
+	if (state.renderData) {
+		if (
+			!state.renderData.locales ||
+			!state.renderData.locales.find(
+				(item) => item.label === currentLocale.label
+			)
+		) {
+			if (state.renderData.defaultLocale) {
+				currentLocale = state.renderData.defaultLocale;
+			}
+			else {
+				currentLocale = defaultLocale;
+			}
+		}
+
+		if (
+			state.renderData.localizedTitles &&
+			state.renderData.localizedTitles[currentLocale.label]
+		) {
+			currentTitle =
+				state.renderData.localizedTitles[currentLocale.label];
+		}
+	}
+
 	const setContentType = (contentType) => {
 		setState({
 			contentType,
@@ -256,13 +395,13 @@ export default ({dataURL, getCache, languageId, spritemap, updateCache}) => {
 				'leftLocalizedContent'
 			)
 		) {
-			if (state.renderData.leftLocalizedContent[languageId]) {
+			if (state.renderData.leftLocalizedContent[currentLocale.label]) {
 				return (
 					<div
 						dangerouslySetInnerHTML={{
 							__html:
 								state.renderData.leftLocalizedContent[
-									languageId
+									currentLocale.label
 								],
 						}}
 					/>
@@ -347,13 +486,13 @@ export default ({dataURL, getCache, languageId, spritemap, updateCache}) => {
 				'rightLocalizedContent'
 			)
 		) {
-			if (state.renderData.rightLocalizedContent[languageId]) {
+			if (state.renderData.rightLocalizedContent[currentLocale.label]) {
 				return (
 					<div
 						dangerouslySetInnerHTML={{
 							__html:
 								state.renderData.rightLocalizedContent[
-									languageId
+									currentLocale.label
 								],
 						}}
 					/>
@@ -429,14 +568,14 @@ export default ({dataURL, getCache, languageId, spritemap, updateCache}) => {
 				'unifiedLocalizedContent'
 			)
 		) {
-			if (state.renderData.unifiedLocalizedContent[languageId]) {
+			if (state.renderData.unifiedLocalizedContent[currentLocale.label]) {
 				return (
 					<div className="taglib-diff-html">
 						<div
 							dangerouslySetInnerHTML={{
 								__html:
 									state.renderData.unifiedLocalizedContent[
-										languageId
+										currentLocale.label
 									],
 							}}
 						/>
@@ -493,6 +632,80 @@ export default ({dataURL, getCache, languageId, spritemap, updateCache}) => {
 		);
 
 		return elements;
+	};
+
+	const renderDropdownMenu = () => {
+		if (!showDropdown || !state.renderData) {
+			return null;
+		}
+
+		const dropdownItems = [];
+
+		if (state.renderData.editURL) {
+			dropdownItems.push({
+				href: state.renderData.editURL,
+				label: Liferay.Language.get('edit'),
+				symbolLeft: 'pencil',
+			});
+		}
+
+		dropdownItems.push({
+			href: discardURL,
+			label: Liferay.Language.get('discard'),
+			symbolLeft: 'times-circle',
+		});
+
+		for (let i = 0; i < dropdownItems.length; i++) {
+			const dropdownItem = dropdownItems[i];
+
+			const href = dropdownItem.href;
+
+			if (typeof href !== 'string') {
+				continue;
+			}
+
+			const index = href.indexOf('?');
+
+			if (index > 0) {
+				let redirectKey = null;
+
+				const params = new URLSearchParams(href.substring(index + 1));
+
+				params.forEach((value, key) => {
+					if (key.endsWith('_redirect')) {
+						redirectKey = key;
+					}
+				});
+
+				if (redirectKey) {
+					params.set(
+						redirectKey,
+						window.location.pathname + window.location.search
+					);
+
+					dropdownItem.href =
+						href.substring(0, index) + '?' + params.toString();
+				}
+			}
+		}
+
+		return (
+			<div className="autofit-col">
+				<ClayDropDownWithItems
+					alignmentPosition={Align.BottomLeft}
+					items={dropdownItems}
+					spritemap={spritemap}
+					trigger={
+						<ClayButtonWithIcon
+							displayType="unstyled"
+							small
+							spritemap={spritemap}
+							symbol="ellipsis-v"
+						/>
+					}
+				/>
+			</div>
+		);
 	};
 
 	const renderViewDropdown = () => {
@@ -627,6 +840,69 @@ export default ({dataURL, getCache, languageId, spritemap, updateCache}) => {
 		);
 	};
 
+	const renderEntry = () => {
+		if (!state.renderData) {
+			if (loading) {
+				return (
+					<div>
+						<span
+							aria-hidden="true"
+							className="loading-animation"
+						/>
+					</div>
+				);
+			}
+
+			return '';
+		}
+		else if (
+			!state.renderData.changeType ||
+			state.renderData.errorMessage
+		) {
+			return (
+				<ClayAlert
+					displayType="danger"
+					spritemap={spritemap}
+					title={Liferay.Language.get('error')}
+				>
+					{state.renderData.errorMessage
+						? state.renderData.errorMessage
+						: Liferay.Language.get('an-unexpected-error-occurred')}
+				</ClayAlert>
+			);
+		}
+
+		return (
+			<table className="publications-render-view table">
+				{renderToolbar()}
+
+				{renderDividers()}
+
+				<tr>
+					{(state.view === VIEW_LEFT ||
+						state.view === VIEW_SPLIT) && (
+						<td className="publications-render-view-content">
+							{renderContentLeft()}
+						</td>
+					)}
+
+					{(state.view === VIEW_RIGHT ||
+						state.view === VIEW_SPLIT) && (
+						<td className="publications-render-view-content">
+							{renderContentRight()}
+						</td>
+					)}
+
+					{state.view === VIEW_UNIFIED && (
+						<td className="publications-render-view-content">
+							{renderContentUnified()}
+						</td>
+					)}
+				</tr>
+			</table>
+		);
+	};
+
 	const renderToolbar = () => {
 		if (state.renderData.changeType === CHANGE_TYPE_PRODUCTION) {
 			return '';
@@ -639,7 +915,7 @@ export default ({dataURL, getCache, languageId, spritemap, updateCache}) => {
 		}
 
 		return (
-			<tr className={loading ? 'publications-loading' : ''}>
+			<tr>
 				<td
 					className="publications-render-view-toolbar"
 					colSpan={columns}
@@ -732,56 +1008,29 @@ export default ({dataURL, getCache, languageId, spritemap, updateCache}) => {
 		);
 	};
 
-	if (!state.renderData) {
-		if (loading) {
-			return (
-				<div>
-					<span aria-hidden="true" className="loading-animation" />
-				</div>
-			);
-		}
-
-		return '';
-	}
-	else if (!state.renderData.changeType || state.renderData.errorMessage) {
-		return (
-			<ClayAlert
-				displayType="danger"
-				spritemap={spritemap}
-				title={Liferay.Language.get('error')}
-			>
-				{state.renderData.errorMessage
-					? state.renderData.errorMessage
-					: Liferay.Language.get('an-unexpected-error-occurred')}
-			</ClayAlert>
-		);
-	}
-
 	return (
-		<table className="publications-render-view table">
-			{renderToolbar()}
+		<div className={`sheet ${loading ? 'publications-loading' : ''}`}>
+			{state.renderData && (
+				<div className="autofit-row sheet-title">
+					{state.renderData.locales &&
+						state.renderData.locales.length > 0 && (
+							<LocalizationDropdown
+								currentLocale={currentLocale}
+								defaultLocale={state.renderData.defaultLocale}
+								locales={state.renderData.locales}
+								setSelectedLocale={setSelectedLocale}
+								spritemap={spritemap}
+							/>
+						)}
+					<div className="autofit-col autofit-col-expand">
+						<h2>{currentTitle}</h2>
 
-			{renderDividers()}
-
-			<tr className={loading ? 'publications-loading' : ''}>
-				{(state.view === VIEW_LEFT || state.view === VIEW_SPLIT) && (
-					<td className="publications-render-view-content">
-						{renderContentLeft()}
-					</td>
-				)}
-
-				{(state.view === VIEW_RIGHT || state.view === VIEW_SPLIT) && (
-					<td className="publications-render-view-content">
-						{renderContentRight()}
-					</td>
-				)}
-
-				{state.view === VIEW_UNIFIED && (
-					<td className="publications-render-view-content">
-						{renderContentUnified()}
-					</td>
-				)}
-			</tr>
-		</table>
+						<div className="entry-description">{description}</div>
+					</div>
+					{renderDropdownMenu()}
+				</div>
+			)}
+			<div className="sheet-section">{renderEntry()}</div>
+		</div>
 	);
 };
